demo/pyproject.toml
[tool.poetry]
name = "demo"
version = "0.1.0"
description = ""
authors = ["MaxHolly <150594066+MaxHolly@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
pandas = "^2.2.1"
ipykernel = "^6.29.3"
jupyter = "^1.0.0"
seaborn = "^0.13.2"
flask = "^3.0.2"
flask-sqlalchemy = "^3.1.1"
werkzeug = "^3.0.1"
flask-wtf = "^1.2.1"
flask-login = "^0.6.3"
imageio = "^2.34.1"
bleach = "^6.1.0"
alembic = "^1.13.2"
sqlalchemy = "^2.0.31"
flask-jwt-extended = "^4.6.0"
flask-migrate = "^4.0.7"
mysqlclient = "^2.2.4"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
